requirements.txt
boto3==1.18.33
botocore==1.21.33
keras==2.6.0
Keras-Preprocessing==1.1.2
numpy==1.19.5
Pillow==8.3.1
requests==2.26.0
s3transfer==0.5.0
tensorflow==2.6.0
Flask==2.0.1